Mode 관련 로직 Controller에서 분리
