+++ src/main/java/controller/BookController.java
@@ -5,32 +5,15 @@ import manager.console.OutputManager;
 import service.BookService;
 
 public class BookController {
-    private BookService bookService;
-    private final OutputManager outputManager = new OutputManager();
-    private final InputManager inputManager = new InputManager();
+    private final BookService bookService;
+    private final InputManager inputManager;
+    private final OutputManager outputManager;
 
-    public void selectMode() {
-        outputManager.printSelectMode();
-        String mode;
-        try {
-            mode = inputManager.getInput();
-        } catch (Exception e) {
-            outputManager.printSystem(e.getMessage());
-            return;
-        }
-        applyMode(mode);
-    }
 
-    public void applyMode(String mode) {
-        ModeType modeType = ModeType.findModeTypeByMode(mode);
-        if (modeType == null) {
-            outputManager.printSystem("잘못된 입력입니다.");
-            return;
-        }
-        bookService = ModeType.findService(modeType, outputManager);
-        ModeType.printModeExecution(modeType, outputManager);
-
-        selectMenu();
+    public BookController(BookService bookService, InputManager inputManager, OutputManager outputManager) {
+        this.bookService = bookService;
+        this.inputManager = inputManager;
+        this.outputManager = outputManager;
     }
 
     public void selectMenu() {
@@ -55,41 +38,41 @@ public class BookController {
         }
     }
 
-    private void reportLostBook() throws Exception{
+    protected void reportLostBook() throws Exception{
         outputManager.printSystem("도서 분실 처리 메뉴로 넘어갑니다.");
         outputManager.printQuestion("분실 처리할 도서번호를 입력하세요.");
         bookService.reportLostBook(Integer.valueOf(inputManager.getInput()));
         outputManager.printSystem("도서가 분실 처리 되었습니다.");
     }
 
-    private void returnBook() throws Exception {
+    protected void returnBook() throws Exception {
         outputManager.printSystem("도서 반납 메뉴로 넘어갑니다.");
         outputManager.printQuestion("반납할 도서번호를 입력하세요.");
         bookService.returnBook(Integer.valueOf(inputManager.getInput()));
         outputManager.printSystem("도서가 반납 처리 되었습니다.");
     }
 
-    private void borrowBook() throws Exception {
+    protected void borrowBook() throws Exception {
         outputManager.printSystem("도서 대여 메뉴로 넘어갑니다.");
         outputManager.printQuestion("대여할 도서번호를 입력하세요.");
         bookService.borrowBook(Integer.valueOf(inputManager.getInput()));
         outputManager.printSystem("도서가 대여 처리 되었습니다.");
     }
 
-    private void searchBook() throws Exception {
+    protected void searchBook() throws Exception {
         outputManager.printSystem("제목으로 도서 검색 메뉴로 넘어갑니다.");
         outputManager.printQuestion("검색할 도서 제목 일부를 입력하세요.");
         bookService.findBookByTitle(inputManager.getInput());
         outputManager.printSystem("검색된 도서 끝");
     }
 
-    private void showBookList() {
+    protected void showBookList() {
         outputManager.printSystem("전체 도서 목록입니다.");
         bookService.showBookList();
         outputManager.printSystem("도서 목록 끝");
     }
 
-    private void saveBook() throws Exception {
+    protected void saveBook() throws Exception {
         outputManager.printSystem("도서 등록 메뉴로 넘어갑니다.");
         outputManager.printQuestion("등록할 도서 제목을 입력하세요.");
         String title = inputManager.getInput();
@@ -105,7 +88,7 @@ public class BookController {
         outputManager.printSystem("도서 등록이 완료되었습니다.");
     }
 
-    private void deleteBook() throws Exception {
+    protected void deleteBook() throws Exception {
         outputManager.printSystem("도서 삭제 처리 메뉴로 넘어갑니다.");
         outputManager.printQuestion("삭제 처리할 도서번호를 입력하세요.");
         Integer id = Integer.valueOf(inputManager.getInput());
